+++ src/bridge/BridgeSettlerBase.sol
@@ -14,6 +14,7 @@ import {FastDeployer} from "../deployer/FastDeployer.sol";
 import {Basic} from "../core/Basic.sol";
 import {Relay} from "../core/Relay.sol";
 import {LayerZeroOFT} from "../core/LayerZeroOFT.sol";
+import {Underpayment} from "../core/SettlerErrors.sol";
 
 abstract contract BridgeSettlerBase is Basic, Relay, LayerZeroOFT {
     using SafeTransferLib for IERC20;
@@ -89,6 +90,16 @@ abstract contract BridgeSettlerBase is Basic, Relay, LayerZeroOFT {
         } else if (action == uint32(IBridgeSettlerActions.BRIDGE_TO_LAYER_ZERO_OFT.selector)) {
             (IERC20 token, uint256 nativeFee, address oft, bytes memory sendData) = abi.decode(data, (IERC20, uint256, address, bytes));
             bridgeLayerZeroOFT(token, nativeFee, oft, sendData);
+        } else if (action == uint32(IBridgeSettlerActions.UNDERPAYMENT_CHECK.selector)) {
+            (uint256 msgValueMin) = abi.decode(data, (uint256));
+            if (msg.value < msgValueMin) {
+                assembly ("memory-safe") {
+                    mstore(0x00, 0xd17e444b) // selector for `Underpayment(uint256,uint256)`
+                    mstore(0x20, msgValueMin)
+                    mstore(0x40, callvalue())
+                    revert(0x1c, 0x44)
+                }
+            }
         } else {
             return false;
         }